Force ansible 1.9
Newer versions of ansible don't like the usage of sudo in some of the playbooks. Ansible introduced a new feature, "become" (http://docs.ansible.com/ansible/become.html) that is supposed to replace sudo and it is supposed to be backwards compatible, but that's not *my* experience and there is no time to re-write all the playbooks. So ansible 1.9 is what we'll stick with for now.